+++ src/perfTest/groovy/com/monits/gradle/sca/performance/categories/AndroidScenario.groovy
@@ -13,8 +13,11 @@
  */
 package com.monits.gradle.sca.performance.categories
 
+import groovy.transform.CompileStatic
+
 /**
  * A category for Android scenario being tested.
 */
+@CompileStatic
 interface AndroidScenario {
 }