+++ src/perfTest/groovy/com/monits/gradle/sca/performance/fixture/AbstractPerfTestFixture.groovy
@@ -28,13 +28,13 @@ import spock.util.environment.Jvm
 @CompileStatic
 abstract class AbstractPerfTestFixture extends Specification {
     // A sample of gradle versions to be considered in general testing - don't test anything below 2.8, it spams stdout
-    static final List<String> TESTED_GRADLE_VERSIONS = ['2.14.1', '3.5.1', GradleVersion.current().version]
+    static final List<String> TESTED_GRADLE_VERSIONS = ['3.5.1', '4.10', GradleVersion.current().version]
         .asImmutable()
     @SuppressWarnings(['DuplicateStringLiteral', 'UnnecessaryCast'])
     static final List<String> TESTED_GRADLE_VERSIONS_FOR_ANDROID = (['2.14.1', '3.5.1'] +
         (Jvm.current.java8Compatible ? ['4.10', GradleVersion.current().version] : [] as List<String>))
         .takeRight(2).asImmutable()
-    static final String BASELINE_PLUGIN_VERSION = '"com.monits:static-code-analysis-plugin:2.6.9"'
+    static final String BASELINE_PLUGIN_VERSION = '"com.monits:static-code-analysis-plugin:2.6.11"'
 
     static final int NUMBER_OF_CLASSES_TO_ANALYZE = 100